Moved wrapper classes to a single wrapper header file. Changed ElevatorUser to User.
