--- BuildingWrapper.cpp
@@ -1,8 +1,0 @@
-#include "BuildingWrapper.h"
-
-BuildingWrapper::BuildingWrapper(Building* b) : pointerToBuilding(b) {}
-
-BuildingWrapper::~BuildingWrapper()
-{
-    delete pointerToBuilding;
-}
