--- BuildingWrapper.h
@@ -1,8 +1,0 @@
-#include "Building.h"
-
-struct BuildingWrapper
-{
-    Building* pointerToBuilding = nullptr;
-    BuildingWrapper(Building* b);
-    ~BuildingWrapper();
-};
