--- DrillWrapper.cpp
@@ -1,9 +1,0 @@
-#include "DrillWrapper.h"
-
-DrillWrapper::DrillWrapper(Drill* d) : pointerToDrill(d) {}
-
-DrillWrapper::~DrillWrapper()
-{
-    delete pointerToDrill;
-}
-
