--- DrillWrapper.h
@@ -1,8 +1,0 @@
-#include "Drill.h"
-
-struct DrillWrapper
-{
-    Drill* pointerToDrill = nullptr;
-    DrillWrapper(Drill* d);
-    ~DrillWrapper();
-};
