--- ElevatorWrapper.cpp
@@ -1,9 +1,0 @@
-#include "ElevatorWrapper.h"
-
-ElevatorWrapper::ElevatorWrapper(Elevator* ev) : pointerToEV(ev) {}
-
-ElevatorWrapper::~ElevatorWrapper()
-{
-    delete pointerToEV;
-}
-
