--- ElevatorWrapper.h
@@ -1,8 +1,0 @@
-#include "Elevator.h"
-
-struct ElevatorWrapper
-{
-    Elevator* pointerToEV = nullptr;
-    ElevatorWrapper(Elevator* ev);
-    ~ElevatorWrapper();
-};
