--- PencilCaseWrapper.cpp
@@ -1,9 +1,0 @@
-#include "PencilCaseWrapper.h"
-
-PencilCaseWrapper::PencilCaseWrapper(PencilCase* pc) : pointerToPC(pc) {}
-
-PencilCaseWrapper::~PencilCaseWrapper()
-{
-    delete pointerToPC;
-}
-
