--- PencilCaseWrapper.h
@@ -1,8 +1,0 @@
-#include "PencilCase.h"
-
-struct PencilCaseWrapper
-{
-    PencilCase* pointerToPC = nullptr;
-    PencilCaseWrapper(PencilCase* pc);
-    ~PencilCaseWrapper();
-};
