--- SchoolBagWrapper.cpp
@@ -1,8 +1,0 @@
-#include "SchoolBagWrapper.h"
-
-SchoolBagWrapper::SchoolBagWrapper(SchoolBag* sb) : pointerToSB(sb) {}
-
-SchoolBagWrapper::~SchoolBagWrapper()
-{
-    delete pointerToSB;
-}
