--- SchoolBagWrapper.h
@@ -1,8 +1,0 @@
-#include "SchoolBag.h"
-
-struct SchoolBagWrapper
-{
-    SchoolBag* pointerToSB = nullptr;
-    SchoolBagWrapper(SchoolBag* sb);
-    ~SchoolBagWrapper();
-};
